LPS-64676 When no country is specified, override canonical language
When country variants for a language are supported (e.g es_ES, es_AR),
but the bare language itself is not listed in `portal.properties`, and a
hook tries to customize that bare language, let the system customize the
canonical language (i.e. the first one listed in the `locales`
property).