LPS-65718 Ignore unique constraint on upgrade
Previously, some portlets would save service preferences at the level of
layouts. This upgrade process tries to convert them to group level
settings, but it doesn't take into account the case where more than one
instance of that portlet exists in the same group.

The only portable and efficient way to handle this is to ignore the
errors and continue.