LPS-72250 Save related model UUID in discussion element
When exporting/importing a comment as part of the export/import
process of its "owner" entity (e.g. a FileEntry), the correct related
classPK will be available in the newPrimaryKeysMap. But, if we are
publishing a standalone comment that information will be missing.

Save the related model UUID in the discussion element so that later we
can try to recover it when importing the discussion.

Note that here I'm making a lot of assumptions, the most important one
being that the model must be both a PersistedModel and a
StagedModel (i.e. this will only work for service builder
entities). Also, I'm assuming that the "classPK" in the discussion is
the same as the entity's primaryKey -- which may not be true.

So, this will fix some of the common cases, but I fear this cannot
really be solved without introducing some major, non backporteable,
changes. In particular, we'd need a generic way to retrieve an element
via its className/classPK pair, and ensure that all className/classPK
pairs used in the portal are consistent with that.