LPS-72546 Add interface for dynamically scoped variables
In portal, we simulate dynamically scoped variables with thread local
variables. The current pattern is to hide the thread local var inside
a util class, and expose methods to get/set the value.

The problem with this approach is twofold:

- It leads to boilerplate. In general, when assigning a value to these
  variables, we want the new value to be valid for a particular
  scope. This requires us to keep the old value, assign the new one,
  and restore the old value in a finally clause. This kind of code
  is duplicated in a lot of places in the portal.

- It is error prone. As all the code for setting/restoring values has
  to be written manually every single time, it is easy to make
  mistakes.

The `DynamicVariable` tries to solve this problem by hiding the
set/reset logic inside it. Callers only need to specify the value, and
provide a function to be executed.