LPS-108987 Avoid gap between sidebar and upper toolbars
This is another example of a discrepancy between `<MultiPanelSidebar />`
and the original sidebar from layout-content-page-editor: the original
one is is sensitive to changes in control menu height on desktop (64px)
vs mobile (56px), but the copy-pasted version was edited to have both be
64px.

This broke the app-builder layout, creating a gap between the toolbar
and the sidebar.

In this commit I apply the minimal fix to force things into sync, but
note that it is not a perfect fix:

- At small screen sizes journal-web has some gaps.

On the bright side, app-builder has no gaps anymore, not even at small
screen sizes (it did have gaps at small screen sizes on master).

The root cause of all this is not just that we have a bunch of different
places where we specify dimensions (and not even always in the same
units) but that we do so variably at different breakpoints, and that we
don't even use breakpoints consistently (we have three ways we do it:
using pixels, using bootstrap "lg" etc breakpoints, and using
"phone"/"tablet" etc which I am not even sure where they are defined).

So this in this PR I take the minimal fix which is to switch to pixels
and force things into sync. It is impossible to fix this 100% without
touching many places in an invasive way, so this is an attempt to make
it mostly better without rewriting the world.

I think the eventual solution for the `<MultiPanelSidebar />` will be
something like this:

- Mount sidebar via a React portal inside the top `position: sticky`
  nav. Put it at the bottom of the nav inside a `position: relative`
  container with zero height.
- Take sidebar out of flow using `position: absolute` and `top: 0`.
- Make nav `overflow: visible` so that we can see the sidebar that is
  actually outside of it.
- Set sidebar height with combination of `calc` and `vh`, and if that
  doesn't work, fallback to doing a JS-based calculation as a last
  resort.