LPS-108921 Let callers choose single or multiple select
Item Selector already has a 'singleSelect' attribute, that signals the
frontend component that the view will return one or multiple results.
Ideally, that same frontend component would send the parameter to the view,
but for the moment, we need to duplicate that info in the Js configuration
and the URL. We'll fix that in a different LPS.

We're keeping the 'multipleSelection' parameter of the Asset Browser
for compatibility reasons. Eventually, we'll remove all Asset Browser
usages and use Item Selector in its place.