LPS-00000 Include container name in error messages
It is possible that files with the same name are provided (some raw
files; some contained inside a ZIP file). This means we cannot use the
file name as a map key. Also, as we need to store multiple properties
instead of a single value, it is better to use a list of maps.