LPS-145232 Always add separator, even if namespace is empty
To avoid problems with field names containing '_', always add an
underscore to separate the namespace from the name, even if the
namespace is empty.

To be able to process ill formed XLF files, accept IDs with no
namespace attached. This way, we'll be able to continue with the
process and display the proper error message.