LPS-203653 Baseline
# breaking

## What modules/apps/adaptive-media/adaptive-media-content-transformer-api/src/main/java/com/liferay/adaptive/media/content/transformer/ContentTransformerHandler.java

The `transform` method has been removed; a new one with no content
type as been added in its place.

## Why

Content types are not used, and all necessary machinery for its
support is a cause for race conditions during portal start up.

----

# breaking

## What modules/apps/adaptive-media/adaptive-media-content-transformer-api/src/main/java/com/liferay/adaptive/media/content/transformer/ContentTransformer.java

The generic type, and the `getContentTransformerContentType` have been
removed.

## Why

Content types are not used, and all necessary machinery for its
support is a cause for race conditions during portal start up.

----